Enhance firmware with logging and configuration for remote log server
- Updated the firmware to include a new log server configuration, allowing users to specify a URL and upload interval for automatic log uploads.
- Added UI elements to the web interface for configuring the log server settings, including validation for URL format and interval range.
- Implemented functionality to display restart count, last reset reason, uptime before restart, and heap status before the last restart in the web interface.
- Refactored EEPROM handling to accommodate new logging features and ensure data integrity during operations.

+++ Clock.h
@@ -7,7 +7,7 @@
 #include <time.h>
 
 extern bool use24HourFormat;
-int formatHourForDisplay(int hour);
+uint8_t formatHourForDisplay(uint8_t hour);
 
 namespace ClockEffect {
   void drawDigit(uint8_t *frame, int digit, uint8_t xOffset, uint8_t yOffset) {
